+++ example/lib/main.dart
@@ -14,7 +14,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.red,
       ),
       home: const MyHomePage(title: 'Simple Chips Input Demo'),
     );
@@ -32,7 +32,17 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   String output = '';
+  String? deletedChip, deletedChipIndex;
   final key = GlobalKey<FormState>();
+  final FocusNode focusNode = FocusNode();
+  final TextFormFieldStyle style = const TextFormFieldStyle(
+    keyboardType: TextInputType.phone,
+    cursorColor: Colors.red,
+    decoration: InputDecoration(
+      contentPadding: EdgeInsets.all(0.0),
+      border: InputBorder.none,
+    ),
+  );
 
   @override
   Widget build(BuildContext context) {
@@ -47,27 +57,39 @@ class _MyHomePageState extends State<MyHomePage> {
           const Padding(
             padding: EdgeInsets.only(bottom: 8.0),
             child: Text(
-              'Example:\nChips with standard email validation',
+              'Example:\nChips with phone number validation',
               textAlign: TextAlign.center,
-              style: TextStyle(fontSize: 20, color: Colors.blue),
+              style: TextStyle(
+                  fontSize: 20,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.black),
             ),
+          ),
+          Text(
+            deletedChip != null
+                ? 'Deleted chip: $deletedChip at index $deletedChipIndex'
+                : '',
+            textAlign: TextAlign.center,
+            style: const TextStyle(fontSize: 16, color: Colors.red),
           ),
           Text(
             'Output:\n$output',
             textAlign: TextAlign.center,
-            style: const TextStyle(fontSize: 16, color: Colors.blue),
+            style: const TextStyle(fontSize: 16, color: Colors.red),
           ),
           Padding(
             padding: const EdgeInsets.all(8.0),
             child: SimpleChipsInput(
               separatorCharacter: ",",
+              focusNode: focusNode,
               validateInput: true,
+              autoFocus: true,
               formKey: key,
+              textFormFieldStyle: style,
               validateInputMethod: (String value) {
-                final emailRegExp = RegExp(
-                    r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+");
-                if (!emailRegExp.hasMatch(value)) {
-                  return 'Please enter a valid email';
+                final phoneNumberRegExp = RegExp(r'(^(?:[+0]9)?[0-9]{10,12}$)');
+                if (!phoneNumberRegExp.hasMatch(value)) {
+                  return 'Enter a valid phone number';
                 }
               },
               onSubmitted: (p0) {
@@ -75,25 +97,38 @@ class _MyHomePageState extends State<MyHomePage> {
                   output = p0;
                 });
               },
+              onChipDeleted: (p0, p1) {
+                setState(() {
+                  deletedChip = p0;
+                  deletedChipIndex = p1.toString();
+                });
+              },
               onSaved: ((p0) {
                 setState(() {
                   output = p0;
                 });
               }),
+              chipTextStyle: const TextStyle(
+                color: Colors.black,
+                fontSize: 16,
+              ),
+              deleteIcon: Icons.delete,
+              deleteIconColor: Colors.black,
               widgetContainerDecoration: BoxDecoration(
-                color: Colors.blueAccent.withOpacity(0.2),
+                color: Colors.redAccent.withOpacity(0.2),
                 borderRadius: BorderRadius.circular(16.0),
-                border: Border.all(color: Colors.blue),
+                border: Border.all(color: Colors.red),
               ),
               chipContainerDecoration: BoxDecoration(
-                color: Colors.blue,
+                color: Colors.yellow,
+                border: Border.all(color: Colors.black),
                 borderRadius: BorderRadius.circular(50),
               ),
               placeChipsSectionAbove: false,
             ),
           ),
           MaterialButton(
-            color: Colors.blue,
+            color: Colors.red,
             onPressed: (() {
               key.currentState!.save();
             }),